+++ bin/lib/cli/cli.prune.ts
@@ -4,11 +4,13 @@ import path from 'path';
 import os from 'os';
 import { File, Folder } from '../../src/filesystem';
 import { handleErrorSync } from '@stlib/utils';
+import { sequelize } from '../db';
 
 export const serverPrune = async (options: OptionValues) => {
   try {
     await clearConfig(options);
     await clearStorage(options);
+    await clearDatabase(options);
     await clearAllData(options);
   } catch (error) {
     handleErrorSync(error, { throw: true });
@@ -29,13 +31,20 @@ const clearStorage = async (options: OptionValues) => {
   }
 };
 
+const clearDatabase = async (options: OptionValues) => {
+  if (options.database) {
+    await sequelize.drop();
+    return console.log('User database has been cleared');
+  }
+};
+
 const clearAllData = async (options: OptionValues) => {
   const rootDir = path.join(os.homedir(), '.lcs-cloud-storage');
   const configDir = path.join(os.homedir(), '.config', 'lcs-cloud-storage');
 
   if (options.all) {
-    await Folder.remove(rootDir).then(() => {
-      Folder.remove(configDir);
+    await Folder.remove(rootDir).then(async () => {
+      await Folder.remove(configDir);
     });
     return console.log('All data has been cleared.');
   }