+++ bin/src/cli/cli.ts
@@ -19,6 +19,10 @@ program
   .option('--dport <dport>', 'define default server port')
   .option('--jwtkey <jwtkey>', 'define jwt key to sign tokens')
   .option('--dbname <dbname>', 'define name for users database')
+  .option(
+    '--authexp <time>',
+    'define a life time for login sessions. Format: "2 days", "10h", "7d"',
+  )
   .action(async (options) => {
     await configure(options);
   });
@@ -39,6 +43,7 @@ program
   .description('Clear server data and delete it')
   .option('-c, --config', 'clear server config')
   .option('-s, --storage', 'clear server storage')
+  .option('-db, --database', 'drop user database')
   .option('-a, --all', 'clear all server data from system')
   .action(async (options) => {
     await serverPrune(options);